Fix widget not updating on first click
This fixes a bug where the widget was not responding to clicks after
it was added for the first time. Only adding another widget resulted
in an upate call after which the first widget would start responding to
clicks.

Looks like the bug was somehow related  to calling
AppWidgetManager.updateAppWidget multiple times from a single
update. As for why it was working after adding a second widget I'm
don't really know.