Suppress compound drawable warning
The image/text combination under LinearLayout produces a warning that suggests
using a TextView with a comound drawable (meaning the drawableStart etc. attribute
on the TextView). Tried doing that but had a hard time scaling the image to the right
size with that approach. Perhaps should just scale the original file (and also have
multiple versions of that for devices with different dpi). Suppressing the warning for
now anyway.